Fixed database bug, added some CSS styling
Fixed a database bug that occured if a user had any '.' characters in
their email address, since SQL uses this character as a schema extender
character, everything before the '.' would be added as a new schema and
everything after was the table name. Also, added some CSS styling and
added a "Register" button to the login page.